refactor: move common components into subfolder
