+++ src/components/common/layout.tsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+import { Flex } from "@rebass/grid";
+
+const Card = styled(Flex)`
+    background-color: #0e1a49;
+    border-radius: 2px;
+    box-shadow: 0 0 12px 6px rgba(155, 155, 155, 0.045);
+    width: 100%;
+    padding: 15px 20px;
+    min-height: 100px;
+`;
+
+export { Card };
